BalajiN , Shruthi | #59 (WHP) | Refactoring
	1. Ektorp supports only int pageNumbers
	2. Implementing pagination in the AdherenceService

+++ motech-reports/src/main/java/org/motechproject/reports/builder/PagedReportBuilder.java
@@ -32,5 +32,5 @@ public abstract class PagedReportBuilder<T> {
 
     protected abstract List<String> createRowData(T modal);
 
-    protected abstract List<T> data(long pageNumber);
+    protected abstract List<T> data(int pageNumber);
 }
